@@ -13,21 +13,20 @@
   },
   "license": "MIT",
   "dependencies": {
-    "@angular/common": "^4.0.0",
-    "@angular/compiler": "^4.0.0",
-    "@angular/core": "^4.0.0",
-    "@angular/forms": "^4.0.0",
-    "@angular/http": "^4.0.0",
-    "@angular/platform-browser": "^4.0.0",
-    "@angular/platform-browser-dynamic": "^4.0.0",
-    "@angular/router": "^4.0.0",
+    "@angular/common": "^5.0.0",
+    "@angular/compiler": "^5.0.0",
+    "@angular/core": "^5.0.0",
+    "@angular/forms": "^5.0.0",
+    "@angular/platform-browser": "^5.0.0",
+    "@angular/platform-browser-dynamic": "^5.0.0",
+    "@angular/router": "^5.0.0",
     "core-js": "^2.4.1",
     "rxjs": "^5.2.0",
-    "systemjs": "^0.19.47",
+    "systemjs": "^0.20.19",
     "zone.js": "^0.8.5"
   },
   "devDependencies": {
-    "@types/node": "^6.0.60",
+    "@types/node": "^8.0.53",
     "concurrently": "^3.1.0",
     "lite-server": "^2.3.0",
     "typescript": "^2.2.2"
