@@ -1,88 +1,87 @@
-﻿import { Http, BaseRequestOptions, Response, ResponseOptions, RequestMethod, XHRBackend, RequestOptions } from '@angular/http';
-import { MockBackend, MockConnection } from '@angular/http/testing';
+﻿import { Injectable } from '@angular/core';
+import { HttpRequest, HttpResponse, HttpHandler, HttpEvent, HttpInterceptor, HTTP_INTERCEPTORS } from '@angular/common/http';
+import { Observable } from 'rxjs/Observable';
+import 'rxjs/add/observable/of';
+import 'rxjs/add/observable/throw';
+import 'rxjs/add/operator/delay';
+import 'rxjs/add/operator/mergeMap';
+import 'rxjs/add/operator/materialize';
+import 'rxjs/add/operator/dematerialize';
 
-export function fakeBackendFactory(backend: MockBackend, options: BaseRequestOptions, realBackend: XHRBackend) {
-    // array in local storage for registered users
-    let users: any[] = JSON.parse(localStorage.getItem('users')) || [];
+@Injectable()
+export class FakeBackendInterceptor implements HttpInterceptor {
 
-    // configure fake backend
-    backend.connections.subscribe((connection: MockConnection) => {
-        // wrap in timeout to simulate server api call
-        setTimeout(() => {
+    constructor() { }
+
+    intercept(request: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
+        // array in local storage for registered users
+        let users: any[] = JSON.parse(localStorage.getItem('users')) || [];
+
+        // wrap in delayed observable to simulate server api call
+        return Observable.of(null).mergeMap(() => {
 
             // authenticate
-            if (connection.request.url.endsWith('/api/authenticate') && connection.request.method === RequestMethod.Post) {
-                // get parameters from post request
-                let params = JSON.parse(connection.request.getBody());
-
+            if (request.url.endsWith('/api/authenticate') && request.method === 'POST') {
                 // find if any user matches login credentials
                 let filteredUsers = users.filter(user => {
-                    return user.username === params.username && user.password === params.password;
+                    return user.username === request.body.username && user.password === request.body.password;
                 });
 
                 if (filteredUsers.length) {
                     // if login details are valid return 200 OK with user details and fake jwt token
                     let user = filteredUsers[0];
-                    connection.mockRespond(new Response(new ResponseOptions({
-                        status: 200,
-                        body: {
-                            id: user.id,
-                            username: user.username,
-                            firstName: user.firstName,
-                            lastName: user.lastName,
-                            token: 'fake-jwt-token'
-                        }
-                    })));
+                    let body = {
+                        id: user.id,
+                        username: user.username,
+                        firstName: user.firstName,
+                        lastName: user.lastName,
+                        token: 'fake-jwt-token'
+                    };
+
+                    return Observable.of(new HttpResponse({ status: 200, body: body }));
                 } else {
                     // else return 400 bad request
-                    connection.mockError(new Error('Username or password is incorrect'));
+                    return Observable.throw('Username or password is incorrect');
                 }
-
-                return;
             }
 
             // get users
-            if (connection.request.url.endsWith('/api/users') && connection.request.method === RequestMethod.Get) {
+            if (request.url.endsWith('/api/users') && request.method === 'GET') {
                 // check for fake auth token in header and return users if valid, this security is implemented server side in a real application
-                if (connection.request.headers.get('Authorization') === 'Bearer fake-jwt-token') {
-                    connection.mockRespond(new Response(new ResponseOptions({ status: 200, body: users })));
+                if (request.headers.get('Authorization') === 'Bearer fake-jwt-token') {
+                    return Observable.of(new HttpResponse({ status: 200, body: users }));
                 } else {
                     // return 401 not authorised if token is null or invalid
-                    connection.mockRespond(new Response(new ResponseOptions({ status: 401 })));
+                    return Observable.throw('Unauthorised');
                 }
-
-                return;
             }
 
             // get user by id
-            if (connection.request.url.match(/\/api\/users\/\d+$/) && connection.request.method === RequestMethod.Get) {
+            if (request.url.match(/\/api\/users\/\d+$/) && request.method === 'GET') {
                 // check for fake auth token in header and return user if valid, this security is implemented server side in a real application
-                if (connection.request.headers.get('Authorization') === 'Bearer fake-jwt-token') {
+                if (request.headers.get('Authorization') === 'Bearer fake-jwt-token') {
                     // find user by id in users array
-                    let urlParts = connection.request.url.split('/');
+                    let urlParts = request.url.split('/');
                     let id = parseInt(urlParts[urlParts.length - 1]);
                     let matchedUsers = users.filter(user => { return user.id === id; });
                     let user = matchedUsers.length ? matchedUsers[0] : null;
 
-                    // respond 200 OK with user
-                    connection.mockRespond(new Response(new ResponseOptions({ status: 200, body: user })));
+                    return Observable.of(new HttpResponse({ status: 200, body: user }));
                 } else {
                     // return 401 not authorised if token is null or invalid
-                    connection.mockRespond(new Response(new ResponseOptions({ status: 401 })));
+                    return Observable.throw('Unauthorised');
                 }
-
-                return;
             }
 
             // create user
-            if (connection.request.url.endsWith('/api/users') && connection.request.method === RequestMethod.Post) {
+            if (request.url.endsWith('/api/users') && request.method === 'POST') {
                 // get new user object from post body
-                let newUser = JSON.parse(connection.request.getBody());
+                let newUser = request.body;
 
                 // validation
                 let duplicateUser = users.filter(user => { return user.username === newUser.username; }).length;
                 if (duplicateUser) {
-                    return connection.mockError(new Error('Username "' + newUser.username + '" is already taken'));
+                    return Observable.throw('Username "' + newUser.username + '" is already taken');
                 }
 
                 // save new user
@@ -91,17 +90,15 @@ export function fakeBackendFactory(backend: MockBackend, options: BaseRequestOpt
                 localStorage.setItem('users', JSON.stringify(users));
 
                 // respond 200 OK
-                connection.mockRespond(new Response(new ResponseOptions({ status: 200 })));
-
-                return;
+                return Observable.of(new HttpResponse({ status: 200 }));
             }
 
             // delete user
-            if (connection.request.url.match(/\/api\/users\/\d+$/) && connection.request.method === RequestMethod.Delete) {
+            if (request.url.match(/\/api\/users\/\d+$/) && request.method === 'DELETE') {
                 // check for fake auth token in header and return user if valid, this security is implemented server side in a real application
-                if (connection.request.headers.get('Authorization') === 'Bearer fake-jwt-token') {
+                if (request.headers.get('Authorization') === 'Bearer fake-jwt-token') {
                     // find user by id in users array
-                    let urlParts = connection.request.url.split('/');
+                    let urlParts = request.url.split('/');
                     let id = parseInt(urlParts[urlParts.length - 1]);
                     for (let i = 0; i < users.length; i++) {
                         let user = users[i];
@@ -114,43 +111,28 @@ export function fakeBackendFactory(backend: MockBackend, options: BaseRequestOpt
                     }
 
                     // respond 200 OK
-                    connection.mockRespond(new Response(new ResponseOptions({ status: 200 })));
+                    return Observable.of(new HttpResponse({ status: 200 }));
                 } else {
                     // return 401 not authorised if token is null or invalid
-                    connection.mockRespond(new Response(new ResponseOptions({ status: 401 })));
+                    return Observable.throw('Unauthorised');
                 }
-
-                return;
             }
 
             // pass through any requests not handled above
-            let realHttp = new Http(realBackend, options);
-            let requestOptions = new RequestOptions({
-                method: connection.request.method,
-                headers: connection.request.headers,
-                body: connection.request.getBody(),
-                url: connection.request.url,
-                withCredentials: connection.request.withCredentials,
-                responseType: connection.request.responseType
-            });
-            realHttp.request(connection.request.url, requestOptions)
-                .subscribe((response: Response) => {
-                    connection.mockRespond(response);
-                },
-                (error: any) => {
-                    connection.mockError(error);
-                });
+            return next.handle(request);
+            
+        })
 
-        }, 500);
-
-    });
-
-    return new Http(backend, options);
-};
+        // call materialize and dematerialize to ensure delay even if an error is thrown (https://github.com/Reactive-Extensions/RxJS/issues/648)
+        .materialize()
+        .delay(500)
+        .dematerialize();
+    }
+}
 
 export let fakeBackendProvider = {
     // use fake backend in place of Http service for backend-less development
-    provide: Http,
-    useFactory: fakeBackendFactory,
-    deps: [MockBackend, BaseRequestOptions, XHRBackend]
+    provide: HTTP_INTERCEPTORS,
+    useClass: FakeBackendInterceptor,
+    multi: true
 };
