@@ -1,1 +1,2 @@
-﻿export * from './fake-backend';
+﻿export * from './jwt.interceptor';
+export * from './fake-backend';
